+++ packages/vite-plugin/src/mpaPlugin.ts
@@ -7,6 +7,7 @@ import { posix } from "node:path";
 import { cwd } from "node:process";
 import { type PioneerPluginOptions } from ".";
 import { RollupOptions } from "rollup";
+import { AdvancedAppOptions } from "../types";
 
 export function mpaPlugin(options: PioneerPluginOptions | undefined): Plugin {
     const rootSite = options?.rootSite ?? false;
@@ -19,17 +20,26 @@ export function mpaPlugin(options: PioneerPluginOptions | undefined): Plugin {
 
         config(config) {
             const sourceRoot = config.root ?? cwd();
-            const entryPoints = gatherEntryPoints({ apps, sites, rootSite, sourceRoot });
+            const { entryPoints, appNames } = gatherEntryPoints({
+                apps,
+                sites,
+                rootSite,
+                sourceRoot
+            });
             const rollupOptions: RollupOptions = {
                 input: entryPoints,
                 output: {
                     entryFileNames(chunk) {
-                        if (apps.includes(chunk.name)) {
+                        if (appNames.includes(chunk.name)) {
                             return "[name].js";
                         }
 
                         // This will rename the .js files that belong to a .html site, they don't need a public name.
-                        return posix.join(resolvedConfig.build.assetsDir, "[name]-[hash].js");
+                        return posix.join(resolvedConfig.build.assetsDir, "[hash:12].js");
+                    },
+                    chunkFileNames(chunk) {
+                        void chunk; // ignored
+                        return posix.join(resolvedConfig.build.assetsDir, "[hash:12].js");
                     }
                 }
             };
@@ -47,20 +57,46 @@ export function mpaPlugin(options: PioneerPluginOptions | undefined): Plugin {
     };
 }
 
+export interface EntryPointInfo {
+    // Parsed rollup entry points
+    entryPoints: Record<string, string>;
+
+    // All app names used by the project
+    appNames: string[];
+}
+
 function gatherEntryPoints(options: {
-    apps: string[];
+    apps: string[] | AdvancedAppOptions;
     sites: string[];
     rootSite: boolean;
     sourceRoot: string;
-}): Record<string, string> {
-    const apps = options.apps.map((appName) => {
-        const prefix = resolve(options.sourceRoot, "apps", appName, "app");
-        const path = findMatchingEntryPointFile(prefix);
-        return {
-            name: appName,
-            path: path
-        };
-    });
+}): EntryPointInfo {
+    interface RawEntryPoint {
+        name: string;
+        path: string;
+    }
+
+    let apps: RawEntryPoint[];
+    if (Array.isArray(options.apps)) {
+        apps = options.apps.map((appName) => {
+            const prefix = resolve(options.sourceRoot, "apps", appName, "app");
+            const path = findMatchingEntryPointFile(appName, prefix);
+            return {
+                name: appName,
+                path: path
+            };
+        });
+    } else {
+        apps = Object.entries(options.apps).map(([name, appPath]) => {
+            const path = resolve(options.sourceRoot, appPath);
+            if (!existsSync(path)) {
+                throw new Error(
+                    `Failed to find app '${name}' at ${path}. Ensure that the path is spelled correctly.`
+                );
+            }
+            return { name, path };
+        });
+    }
 
     /*
      * Vite does not respect the entry point name for html files, it
@@ -70,22 +106,24 @@ function gatherEntryPoints(options: {
      * There are multiple vite mpa plugins that handle custom html paths which can either be used
      * directly or as inspiration.
      */
-    const sites = options.sites.map((siteName) => {
-        const path = resolve(options.sourceRoot, "sites", siteName, "index.html");
+    const sites = options.sites.map((sitePath) => {
+        const path = resolve(options.sourceRoot, sitePath, "index.html");
         if (!existsSync(path)) {
             throw new Error(
-                `Failed to find site '${siteName}' at ${path}. Please ensure that the site name is spelled correctly.`
+                `Failed to find site at ${path}. Ensure that the site name in your vite configuration is spelled correctly.`
             );
         }
         return {
-            name: siteName,
+            name: sitePath,
             path: path
         };
     });
     if (options.rootSite) {
         const path = resolve(options.sourceRoot, "index.html");
         if (!existsSync(path)) {
-            throw new Error(`Failed to find root site at ${path}. Please create the missing file.`);
+            throw new Error(
+                `Failed to find root site at ${path}. Create the missing file or set 'rootSite' to false.`
+            );
         }
         sites.push({
             name: "index",
@@ -99,7 +137,7 @@ function gatherEntryPoints(options: {
         const existingPath = entryPoints[name];
         if (existingPath) {
             throw new Error(
-                `Entry point '${name}' was defined twice (paths ${existingPath} and ${path}). Please remove the duplicated entry.`
+                `Entry point '${name}' was defined twice (paths ${existingPath} and ${path}). Remove the duplicated entry.`
             );
         }
         entryPoints[name] = path;
@@ -110,12 +148,17 @@ function gatherEntryPoints(options: {
             "You must configure at least one site or one app in the pioneer plugin options."
         );
     }
-    return entryPoints;
+
+    const appNames = apps.map(({ name }) => name);
+    return {
+        entryPoints,
+        appNames
+    };
 }
 
 const EXTENSIONS = [".ts", ".tsx", ".js", ".jsx"];
 
-function findMatchingEntryPointFile(prefix: string) {
+function findMatchingEntryPointFile(appName: string, prefix: string) {
     for (const ext of EXTENSIONS) {
         const candidate = prefix + ext;
         if (existsSync(candidate)) {
@@ -125,6 +168,8 @@ function findMatchingEntryPointFile(prefix: string) {
 
     const extensions = EXTENSIONS.join(", ");
     throw new Error(
-        `Failed to find a matching entry point file for '${prefix}'. Supported extensions are ${extensions}. Ensure that the app name is spelled correctly or create the missing file.`
+        `Failed to find a matching entry point file for app '${appName}' at ${prefix}.` +
+            ` Supported extensions are ${extensions}.` +
+            ` Ensure that the app name is spelled correctly or create the missing file.`
     );
 }