+++ packages/vite-plugin/test-data/multi-page-advanced-apps/custom/location/app.ts
@@ -1,0 +1,2 @@
+console.log("hello from custom location");
+export default null;